backend/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-huggingface>=0.0.1  # Midm 모델 사용 시 필요
psycopg2-binary>=2.9.0
pgvector>=0.2.0
numpy>=1.24.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
# Midm 모델 사용 시 추가 필요:
# transformers>=4.39.0
# torch>=2.0.0

